### 2020-02-03 #2064 @whedon
[PRE REVIEW]: `psi-collect`: A Python module for post-storm image collection and cataloging
**Submitting author:** @ebgoldstein (<a href="http://orcid.org/0000-0001-9358-1016">Evan B Goldstein</a>)
**Repository:** <a href="https://github.com/UNCG-DAISY/psi-collect" target ="_blank">https://github.com/UNCG-DAISY/psi-collect</a>
**Version:** v0.6
