### 2020-02-04 #2064 @kbarnhart
@ebgoldstein thanks for recommending reviewers. I will work on getting two or three reviewers for this submission. Once I've gotten a few reviewers, we will close this issue and start the review on a new dedicated review issue.

A couple of comments from me.

1. I would recommend adding a bit more motivation to the paper.md. Based on our conversation at AGU I know that there many qualities of this package that make collection and cataloging of images _way easier_, probably more appropriately described as _actually possible_. This doesn't come through clearly in the paper. I would recommend expanding the statement of need and/or adding a section that focuses on challenges to accessing this dataset that sets up the statement of need.

2. Reference `Madore2018` needs and `@` in front.
