### 2020-02-04 #2064 @ebgoldstein
i corrected the madore ref, and will work on adding more motivation.
