### 2020-02-04 UNCG-DAISY/psi-collect#62 @ebgoldstein
add more detail to paper.md
> I would recommend adding a bit more motivation to the paper.md. Based on our conversation at AGU I know that there many qualities of this package that make collection and cataloging of images way easier, probably more appropriately described as actually possible. This doesn't come through clearly in the paper. I would recommend expanding the statement of need and/or adding a section that focuses on challenges to accessing this dataset that sets up the statement of need.

from:
https://github.com/openjournals/joss-reviews/issues/2064#issuecomment-581974815

@Matmorcat I will add text, but if you have any specific ideas, pls feel free to add or ping me.
