### 2020-02-05 UNCG-DAISY/psi-collect#62 @Matmorcat
@ebgoldstein I'd say to just include elaboration about how the download of archives tended to freeze or terminate regularly throughout attempting to download. There are other software that can resume partial downloads (including Firefox's downloading), but will either not continue when a problem arises that can be solved (Firefox) or they are not able to intelligently handle the actual scraping of the archives for our particular use case (most require you to manually select what you want to download and do not scan multiple levels of pages).

Our program stops at the appropriate depth to capture all of the data archives without unnecessarily scanning further pages, as that would increase the time and bandwidth requirement for finding archives at an increasing rate as you traverse deeper, and would introduce loops (e.g. page A links to page B links to page C links to page A).
