### 2020-02-05 UNCG-DAISY/psi-collect#62 @ebgoldstein
thanks @Matmorcat — your text helps alot. The NGS folks suggested:

`wget -T 3600 --no-http-keep-alive`

but this seems suboptimal as one would still need to supply the address for each image tar or ZIP
