### 2020-02-05 UNCG-DAISY/psi-collect#62 @Matmorcat
Yes, good example. That command will correct for timeouts, but does not traverse pages dynamically (meaning it would not work for the older archives), and it doesn't allow for finding archives for a storm unless you manually select their download URLs.
