### 2020-02-05 UNCG-DAISY/psi-collect#62 @Matmorcat
@ebgoldstein Also, we may have overlooked the importance of the locking mechanism for some teams. In the process of downloading the data for our own uses, we implemented a soft locking mechanism that allowed us to download the archives from multiple computers to our team drive. This could be used with almost any network file system or distributed file systems like Hadoop in which multiple computers would need to download to the same location without stepping on each other's feet.

The locking system creates a text file named the same as the archive, but with the additional ending '.lock' (eg. 'xxxx.tar.lock') that contains information about what user is downloading that archive, their progress, and how large it will be when done downloading (and fully uploaded). It then updates this file regularly as the download progresses for the downloading user.
