+++ README.md
@@ -21,9 +21,9 @@ https://psi-collect.readthedocs.io/en/latest/)
 
 - [**Matthew Moretz**](https://github.com/Matmorcat)
 - [**Daniel Foster**](https://github.com/dlfosterbot)
+- [**Shah Nafis Rafique**](https://github.com/ShahNafisRafique)
 - [**John Weber**](https://github.com/JWeb56)
 - [**Rinty Chowdhury**](https://github.com/rintychy)
-- [**Shah Nafis Rafique**](https://github.com/ShahNafisRafique)
 - [**Evan Goldstein**](https://github.com/ebgoldstein)
 - [**Somya Mohanty**](https://github.com/somyamohanty)
 
