### 2020-02-05 d16aa41
Update development status of program to stable (#65)
