### 2020-02-05 #2075 @whedon
[REVIEW]: `psi-collect`: A Python module for post-storm image collection and cataloging
**Submitting author:** @ebgoldstein (<a href="http://orcid.org/0000-0001-9358-1016">Evan B Goldstein</a>)
**Repository:** <a href="https://github.com/UNCG-DAISY/psi-collect" target ="_blank">https://github.com/UNCG-DAISY/psi-collect</a>
**Version:** v1.0.4

- [x] **Repository:** Is the source code for this software available at the <a
target="_blank" href="https://github.com/UNCG-DAISY/psi-collect">repository url</a>?
- [x] **License:** Does the repository contain a plain-text LICENSE file with
the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license?

- [x] **Community guidelines:** Are there clear guidelines for third parties
wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
