### 2020-02-05 #2075 @kbarnhart
@arfon something funny is happening with whedon and PDF generating. Here and in the PRE-REVIEW (#2064) the automated attempt to compile the paper.md failed, but when I asked whedon to do it, ti succeeded.
