### 2020-02-05 #2075 @arfon
> @arfon something funny is happening with whedon and PDF generating. Here and in the PRE-REVIEW (#2064) the automated attempt to compile the paper.md failed, but when I asked whedon to do it, ti succeeded.

I think the issue here is that the repository is pretty large (> 400Mb) which means that Whedon is sometimes timing out. I'm not sure that there's much we can do about this sorry, unless the authors are able to [reduce the size of the repository](https://www.google.com/search?q=reducing+git+repository+size&oq=reducing+git+repo&aqs=chrome.0.0j69i57j0l4.4260j1j7&sourceid=chrome&ie=UTF-8).
