### 2020-02-05 UNCG-DAISY/psi-collect#62 @ebgoldstein
@Matmorcat,  here is some new text (👇 ). I think we might remove the fourth (?) since its specific to us... (@kbarnhart, patching you in for a look to see if this is what you are after)

👇 👇 👇 👇 👇 👇 👇

`psi-collect` works to address four issues that enable researchers to work with large collections of NOAA post storm aerials.

First, Users previously needed to manually navigate to each storm and individually download the multiple image tar / zip files. Download could be done with the graphical user interface, or via another technique (e.g. `wget` on Linux). Even though `wget` is a command line method for retrieving the data, users would still be required to manually navigate and include individual web addresses for each storm archive (there are 15 individual archives for Hurricane Florence). `psi-collect` dynamically traverses pages and downloads archives when given specific filtering/sub-setting criteria (i.e. downloading all archives from Hurricane Florence)

Second, the individual image archives are large, and often freeze or terminate during download. Some browsers or software tools may be capable of resuming partial downloads, but this is case specific and depends on the retrieval workflow. Terminated downloads may require users to determine the missing archive, and re-initiate a download. `psi-collect` has an automatic resumption feature in the event that a download is interrupted. Additionally, tar and zip archives are checked for the integrity of the contents upon download completion to ensure that the included data are accurate and intact.

Third, we have observed that download speeds can be slow. `psi-collect` handles queuing of downloads, and cataloging what downloads exist on a user's local computer. These are useful features when multiple archives are needed.

Fourth, `psi-collect` implements a soft locking mechanism allowing users to simultaneously download archives on multiple computers that upload to a single network file system or integrate with a distributed file systems (e.g. Hadoop). The locking system creates a text file named the same as the archive, but with an additional ending (e.g. 'xxxx.tar.lock') containing information about what user is downloading that archive, their progress, and how large it will be when done downloading (and fully uploaded). It then updates this file regularly as the download progresses.
