### 2020-02-06 UNCG-DAISY/psi-collect#62 @Matmorcat
I'd say keep the 4th as it is a discrete feature built in, and may be very useful when it comes to surpassing the limit of one computer's network. DFSs are also used for distributed computation w/ Spark and other systems when working with big data, so it might not be unreasonable that some may use such a feature for its ease of use.

@ebgoldstein I made some slight grammar / spelling changes above, you can see the changes by clicking on my username where it says `edited by Matmorcat`
