### 2020-02-07 db3de55
Cache file_origin_size
The size of the archive was not being properly cached and could end up being retrieved multiple times.
