+++ psicollect/collector/archive.py
@@ -38,7 +38,8 @@ class Archive:
     # Save a cache of the file size in bytes
     file_origin_size: int or None = None
 
-    def __init__(self, archive_url: str, archive_date: str = UNKNOWN, archive_label: str = UNKNOWN):
+    def __init__(self, archive_url: str, archive_date: str = UNKNOWN, archive_label: str = UNKNOWN,
+                 file_size_origin: int = None):
         """Initializes the object with required information for a archive file
 
         :param archive_url: The url to download the archive file
@@ -48,6 +49,7 @@ class Archive:
         self.date = archive_date
         self.url = archive_url
         self.type_label = archive_label
+        self.file_origin_size = file_size_origin
 
         # Parse out needed information from the URL
         self.name, self.type = re.findall(re.compile('.*/([^/]+)\\.(tar|zip)', re.IGNORECASE), self.url)[0]
