### 2020-02-07 8d8c5ef
Collection performance improvements (#68)
* Cache file_origin_size

The size of the archive was not being properly cached and could end up being retrieved multiple times.

* Get file sizes concurrently for faster collection

Collection of images can be sped up by getting file sizes concurrently using threading. Currently, sizes are being collected sequentially after the Archive class has been created. We can get large speed increases when creating the archive list by getting all the file sizes first (concurrently), then creating our Archives with these sizes

* Fix formatting

Co-authored-by: 
[…]
