### 2020-02-07 00d6c4d
Update dev installation instructions
