### 2020-02-07 1537614
Bump pymdown-extensions from 6.2.1 to 6.3
Bumps [pymdown-extensions](https://github.com/facelessuser/pymdown-extensions) from 6.2.1 to 6.3.
- [Release notes](https://github.com/facelessuser/pymdown-extensions/releases)
- [Commits](https://github.com/facelessuser/pymdown-extensions/compare/6.2.1...6.3.0)

Signed-off-by: dependabot-preview[bot] <support@dependabot.com>
