### 2020-02-07 4e961c1
Add back docs build status (was left commented out)
