### 2020-02-07 54dfbf2
Normalize formatting for neatness
