### 2020-02-07 ecf9198
Revert "Bump pymdown-extensions from 6.2.1 to 6.3 (#69)"
This reverts commit 4bb8cd9cb7cfba5f02e61aabd31296c956bcef08.
