### 2020-02-07 7e6561b
Clean up hidden badge texts
