### 2020-02-07 cb8043a
Cleanup and fix some issues with the README (#71)
* Update development status of program to stable

* Add back docs build status (was left commented out)

* Normalize formatting for neatness

* Short description probably shouldn't be a header

* Clean up hidden badge texts
