### 2020-02-07 UNCG-DAISY/psi-collect#62 @kbarnhart
@ebgoldstein I think the new text looks good! Thanks for revising.
