### 2020-02-08 UNCG-DAISY/psi-collect#73 @chrisleaman
Cannot specify path for pstorm catalog
**Describe the bug**
`pstorm catalog` expects the downloaded data to only be at `<user_home>/psi/collect/data/archives`. If the user specifies another location to download the files to (`pstorm collect --path C:\data\`), there is no way to specify this path and `pstorm catalog` cannot find the data.

**To Reproduce**
Steps to reproduce the behavior:
1. `pstorm collect -s Florence --path C:\data\ --download`.
2. `pstorm catalog -e tif`

**Expected behavior**
Either one of two options could work:
1. Remove `--path` option from `pstorm collect`. Data will always be downloaded to `<user_home>/psi/collect/data/archives`
2. Add `--path` option to `pstorm catalog` so user can specify where data is located.
