### 2020-02-08 UNCG-DAISY/psi-collect#74 @chrisleaman
Cannot get archive variable from path when making catalog
**Describe the bug**
`pstorm catalog` exits without creating the catalog.

**To Reproduce**
Steps to reproduce the behavior:
1. Download images: `pstorm collect -s Florence -a 20180916b_RGB.tar --download`
2. Run catalog: `pstorm catalog --debug -v 3 -e tif` results in this output (the exit code might indicate some sort of recursion error?)
  <summary>[Output, click to expand]</summary>

```
Parsing out current path to determine catalog variables to use ... Looking for storm in path: C:\Users\z5189959\psi\collect\data\archives
DONE
Getting a list of all valid images ...
Searching through C:\Users\z5189959\psi\collect\data\archives for the pattern ".*" ...
C:\Users\z5189959\psi\collect\data\archives\Florence\20180916b_RGB\20180916bC0793900w325700n.tif ... matches pattern!
C:\Users\z5189959\psi\collect\data\archives\Florence\20180916b_RGB\20180916bC0793900w325830n.tif ... matches pattern!
C:\Users\z5189959\psi\collect\data\archives\Florence\20180916b_RGB\20180916bC0794030w325400n.tif ... matches pattern!
C:\Users\z5189959\psi\collect\data\archives\Florence\20180916b_RGB\20180916bC0794030w325530n.tif ... matches pattern!
C:\Users\z5189959\psi\collect\data\archives\Florence\20180916b_RGB\20180916bC0794030w325700n.tif ... matches pattern!
C:\Users\z5189959\psi\collect\data\archives\Florence\20180916b_RGB\20180916bC0794200w325230n.tif ... matches pattern!
C:\Users\z5189959\psi\collect\data\archives\Florence\20180916b_RGB\20180916bC0794200w325400n.tif ... matches pattern!
C:\Users\z5189959\psi\collect\data\archives\Florence\20180916b_RGB\20180916bC0794200w325530n.tif ... matches pattern!
C:\Users\z5189959\psi\collect\data\archives\Florence\20180916b_RGB\20180916bC0794330w325100n.tif ... matches pattern!
C:\Users\z5189959\psi\collect\data\archives\Florence\20180916b_RGB\20180916bC0794330w325230n.tif ... matches pattern!
C:\Users\z5189959\psi\collect\data\archives\Florence\20180916b_RGB\20180916bC0794330w325400n.tif ... matches pattern!
C:\Users\z5189959\psi\collect\data\archives\Florence\20180916b_RGB\20180916bC0794500w324930n.tif ... matches pattern!
C:\Users\z5189959\psi\collect\data\archives\Florence\20180916b_RGB\20180916bC0794500w325100n.tif ... matches pattern!
C:\Users\z5189959\psi\collect\data\archives\Florence\20180916b_RGB\20180916bC0794500w325230n.tif ... matches pattern!
C:\Users\z5189959\psi\collect\data\archives\Florence\20180916b_RGB\20180916bC0794630w324800n.tif ... matches pattern!
C:\Users\z5189959\psi\collect\data\archives\Florence\20180916b_RGB\20180916bC0794630w324930n.tif ... matches pattern!
C:\Users\z5189959\psi\collect\data\archives\Florence\20180916b_RGB\20180916bC0794800w324800n.tif ... matches pattern!
C:\Users\z5189959\psi\collect\data\archives\Florence\20180916b_RGB\20180916bC0794800w324930n.tif ... matches pattern!
C:\Users\z5189959\psi\collect\data\archives\Florence\20180916b_RGB\20180916bC0794930w324800n.tif ... matches pattern!
C:\Users\z5189959\psi\collect\data\archives\Florence\20180916b_RGB\20180916bC0795100w324630n.tif ... matches pattern!
C:\Users\z5189959\psi\collect\data\archives\Florence\20180916b_RGB\20180916bC0795100w324800n.tif ... matches pattern!
C:\Users\z5189959\psi\collect\data\archives\Florence\20180916b_RGB\20180916bC0795230w324500n.tif ... matches pattern!
C:\Users\z5189959\psi\collect\data\archives\Florence\20180916b_RGB\20180916bC0795230w324630n.tif ... matches pattern!
C:\Users\z5189959\psi\collect\data\archives\Florence\20180916b_RGB\20180916bC0795400w324500n.tif ... matches pattern!
C:\Users\z5189959\psi\collect\data\archives\Florence\20180916b_RGB\20180916bC0795400w324630n.tif ... matches pattern!
 1  Florence/20180916b_RGB/20180916bC0793900w325700n.tif
 2  Florence/20180916b_RGB/20180916bC0793900w325830n.tif
 3  Florence/20180916b_RGB/20180916bC0794030w325400n.tif
 4  Florence/20180916b_RGB/20180916bC0794030w325530n.tif
 5  Florence/20180916b_RGB/20180916bC0794030w325700n.tif
 6  Florence/20180916b_RGB/20180916bC0794200w325230n.tif
 7  Florence/20180916b_RGB/20180916bC0794200w325400n.tif
 8  Florence/20180916b_RGB/20180916bC0794200w325530n.tif
 9  Florence/20180916b_RGB/20180916bC0794330w325100n.tif
10  Florence/20180916b_RGB/20180916bC0794330w325230n.tif
11  Florence/20180916b_RGB/20180916bC0794330w325400n.tif
12  Florence/20180916b_RGB/20180916bC0794500w324930n.tif
13  Florence/20180916b_RGB/20180916bC0794500w325100n.tif
14  Florence/20180916b_RGB/20180916bC0794500w325230n.tif
15  Florence/20180916b_RGB/20180916bC0794630w324800n.tif
16  Florence/20180916b_RGB/20180916bC0794630w324930n.tif
17  Florence/20180916b_RGB/20180916bC0794800w324800n.tif
18  Florence/20180916b_RGB/20180916bC0794800w324930n.tif
19  Florence/20180916b_RGB/20180916bC0794930w324800n.tif
20  Florence/20180916b_RGB/20180916bC0795100w324630n.tif
21  Florence/20180916b_RGB/20180916bC0795100w324800n.tif
22  Florence/20180916b_RGB/20180916bC0795230w324500n.tif
23  Florence/20180916b_RGB/20180916bC0795230w324630n.tif
24  Florence/20180916b_RGB/20180916bC0795400w324500n.tif
25  Florence/20180916b_RGB/20180916bC0795400w324630n.tif
DONE
Parsing out information about images from their paths ...
Process finished with exit code -1073741571 (0xC00000FD)
```

**Expected behavior**
`pstorm catalog` should generate a catalog `.csv`

**Additional context**
The problem appears to be with the `Cataloging._get_archive_from_path` function. If this is disabled, the catalog gets generated:

https://github.com/UNCG-DAISY/psi-collect/blob/cb8043ac06bfd8d028ca18b4921f23b64405fd65/psicollect/cataloging/make_catalog.py#L419-L437

I'm not entirely sure, but I think it might have to do with comparing paths when doing the recursion, because I'm on windows. If I query these variables, I get slightly different paths (note the slashes), due to how `s.DATA_PATH` is constructed. I think paths should be compared using `os.path.normpath` or `os.path.realpath` rather than just using the strings to avoid this issue.

```
In[2]: scope_path
Out[3]: 'C:\\Users\\z5189959\\psi\\collect\\data\\archives\\Florence\\20180916b_RGB\\20180916bC0793900w325700n.tif'
In[3]: s.DATA_PATH
Out[4]: 'C:\\Users\\z5189959/psi/collect\\data'
```
