### 2020-02-08 068014f
Add minor README.md changes (#75)
* Update dev installation instructions

* Update statement of need
