### 2020-02-08 d9acc80
Fix infinite recursion (issue #74) due to wrong split point
