### 2020-02-08 75f4d9a
Fix tif archives (ending in '_RGB') being excluded
