### 2020-02-08 07e88c9
Fix issue with catalogs not being created for TIF archives (#76)
* Fix infinite recursion (issue #74) due to wrong split point

* Fix tif archives (ending in '_RGB') being excluded
