+++ psicollect/cataloging/make_catalog.py
@@ -422,12 +422,12 @@ class Cataloging:
 
         path_head, path_tail = os.path.split(scope_path)
 
-        if path_head == os.path.splitdrive(scope_path)[1] and path_tail == '':
+        if path_tail == '':
             # If the filesystem root directory is reached, a storm-specific catalog cannot be found
 
             raise PathParsingException(objective='the archive name')
 
-        if ('20' in path_tail and '_jpgs' in path_tail) is False or scope_path == s.DATA_PATH:
+        if ('20' in path_tail and '_' in path_tail) is False or scope_path == s.DATA_PATH:
             # If the current directory does not look like an archive name or is the data path
 
             # Keep recursively checking each directory to match the pattern (traverse back through path)
