### 2020-02-08 UNCG-DAISY/psi-collect#74 @Matmorcat
Thank you, it seems the issue was both that TIF archives were not accounted for correctly in the method you linked (required '_jpg' in archive name) and that the split function was not compared correctly and was in fact unnecessary, resulting in infinite recursion!
