+++ .bumpversion.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 1.0.0
+current_version = 1.0.1
 commit = True
 tag = True
 
@@ -26,4 +26,3 @@ test = pytest
 
 [tool:pytest]
 collect_ignore = ['setup.py']
-
