### 2020-02-08 c43d673
Document the current release (v1.0.0+) history of the program
