+++ HISTORY.md
@@ -1,0 +1,10 @@
+# Release Notes
+
+## v1.0.0
+
+- The first stable release
+
+## v1.0.1
+
+- Collection performance improvements (#68) 
+- Fix issue with catalogs not being created for TIF archive (#76) - Fixes Issue #74
