### 2020-02-08 UNCG-DAISY/psi-collect#73 @Matmorcat
I've added the `--path` as an alternative way to tell it to catalog starting in a specific path. The way it works is that it uses the directory the command is run from as the source for the data, but the documentation did not explain that!

I've updated the documentation to reflect this information.
