### 2020-02-08 0e03d50
Add --path to catalog, clarify docs, & Bump 1.0.0 → 1.0.1 (#77)
* Bump version: 1.0.0 → 1.0.1

* Document the current release (v1.0.0+) history of the program

* Add param to override scope path for cataloging

* Add help upon completion of downloading to help w/ cataloging

* Update HISTORY.md

* Misc IDE file change

* Add the --path argument to the docs
