+++ HISTORY.md
@@ -1,0 +1,11 @@
+# Release Notes
+
+## v1.0.0
+
+- The first stable release
+
+## v1.0.1
+
+- Collection performance improvements
+- Fix issue with catalogs not being created for TIF archive (Fixes #74)
+- Make selecting a path for cataloging clearer
