+++ mkdocs.yml
@@ -59,4 +59,4 @@ markdown_extensions:
     - pymdownx.tilde
 
 extra:
-    version: 1.0.0
+    version: 1.0.1
