### 2020-02-08 58c73ed
Improve docs & add help on distribution (#78)
* Bump version: 1.0.0 → 1.0.1

* Document the current release (v1.0.0+) history of the program

* Add param to override scope path for cataloging

* Add help upon completion of downloading to help w/ cataloging

* Update HISTORY.md

* Misc IDE file change

* Add the --path argument to the docs

* Add documentation for drafting new releases for PyPI

* Make sure to push the new version tag

* Group doc pages by purpose
