### 2020-02-08 047daa9
Default docs link to master, not latest, as it is generally more reliable
