### 2020-02-08 a51e1bc
Default docs link to master, not latest, as it is generally more reliable (#79)
