### 2020-02-08 b728877
slight changes to short package desc. in readme.md
