### 2020-02-09 dac3f81
Delete unused comments
