+++ .codecov.yml
@@ -20,8 +20,3 @@ coverage:
   range: 50..90
   round: down
   precision: 2
-#  status:
-#    project:
-#      default:
-#        # basic
-#        target: 75%
