+++ tox.ini
@@ -11,7 +11,7 @@ deps =
     -r {toxinidir}/requirements-dev.txt
 commands =
     pip install -U pip
-    py.test --basetemp={toxinidir} --cov={toxinidir} --cov-report=xml
+    py.test --basetemp={toxinidir}/psicollect --cov={toxinidir}/psicollect --cov-report=xml
 
 passenv =
     CI
