+++ psicollect/.codecov.yml
@@ -4,18 +4,6 @@ ignore:
 #  - "**/*.py"         # glob accepted
 #  - "[a-z]+/test_.*"  # regexp accepted
 
-  # Setup configuration
-  - 'setup.py'
-
-  # Test files
-  - '**/tests'
-
-  # Command-line runners
-  - '**/psicollect/common/pstorm.py'
-  - '**/psicollect/cataloging/catalog.py'
-  - '**/psicollect/cataloging/stack.py'
-  - '**/psicollect/collector/collect.py'
-
 coverage:
   range: 50..90
   round: down
