### 2020-02-09 aa0aea5
Add entry for v1.0.2 to HISTORY.md
