+++ HISTORY.md
@@ -9,3 +9,8 @@
 - Collection performance improvements
 - Fix issue with catalogs not being created for TIF archive (Fixes #74)
 - Make selecting a path for cataloging clearer
+
+## v1.0.2
+
+- Fix issue with some older storms showing no archives
+- Fix display of some paths in collector showing incorrect separator
