### 2020-02-09 12a80e4
Fix older archives no longer showing for older storms & Bump 1.0.1 → 1.0.2 (#80)
* Fix bug with some older storms no longer showing archives

* Fix bug with some older storms no longer showing archives

* Misc project file changes

* Use join method for paths in s.py for consistent display to user

* Add entry for v1.0.2 to HISTORY.md

* Bump version: 1.0.1 → 1.0.2
