### 2020-02-10 UNCG-DAISY/psi-collect#62 @chrisleaman
This added detail looks good 👍 Can you merge into the paper and I'll finish up my review?
