### 2020-02-11 6aa14d2
Bump markdown from 3.1.1 to 3.2 (#82)
Bumps [markdown](https://github.com/Python-Markdown/markdown) from 3.1.1 to 3.2.
- [Release notes](https://github.com/Python-Markdown/markdown/releases)
- [Commits](https://github.com/Python-Markdown/markdown/compare/3.1.1...3.2)

Signed-off-by: dependabot-preview[bot] <support@dependabot.com>

Co-authored-by: Matmorcat <mcmoretz@uncg.edu>
