### 2020-02-11 d614cdd
added more desc. to paper.md
