### 2020-02-11 752bd81
Minor changes to paper.md
