### 2020-02-11 #2075 @chrisleaman
Hi @kbarnhart and `psi-collect` team,

Just letting you know I've finished my review. Overall, I thought it was a nifty tool to manage a lot of unwieldy but important data. I've been finding aerial imagery is so important and can give a lot of context to pre- and post-storm LIDAR surveys especially for assigning storm impacts.

For a bit of feedback on the code, what I liked:
1. The extensive use of type hinting for variables! There were a couple of times I was playing round with the code and it caught a couple of my errors.
2. The fact it's available as a CLI script. Users will find it easy to just call `pstorm` from the CLI, rather than having to implement the package in their own script.
3. Great that tests are included, makes it really easy to see if changes to code affect anything.

These are some suggestions for future improvements, if you're looking for ideas. Some of these I'm sure you're already aware of:

1. Using the `click` package can help tidy up some of the path wrangling that's being done in the `pstorm.py` file. Otherwise perhaps look at the argparse `add_subparsers` command. I'm not sure that using `exec()` to call the script is the best idea.
2. Try reducing the size data in the tests folder. Do you really need to include the full resolution images to test on? Or is it possible to use a reduced size image or even mock them out completely. This might improve compatibility with build systems (which can have issues if the repo is too big) and might just generally be easier to work with.
3. I think a lot of the Path manipulation can be done with the `pathlib` library that was introduced in Python 3.4. This might help clean up the code, rather than using `os.path`.

List of pull requests and issues I submitted:

- https://github.com/UNCG-DAISY/psi-collect/issues/73
- https://github.com/UNCG-DAISY/psi-collect/issues/74
- https://github.com/UNCG-DAISY/psi-collect/pull/75
- https://github.com/UNCG-DAISY/psi-collect/pull/68
- https://github.com/UNCG-DAISY/psi-collect/pull/83
- https://github.com/UNCG-DAISY/psi-collect/pull/84
