### 2020-02-11 704a1ee
Fix typo when calling __main__ (#83)
Co-authored-by: kodiakhq[bot] <49736102+kodiakhq[bot]@users.noreply.github.com>
