+++ runners/pstorm.py
@@ -28,5 +28,5 @@ def main():
               % (s.ROOT_CMD, '|'.join(PATHS)))
 
 
-if __name__ == 'main':
+if __name__ == '__main__':
     main()
