### 2020-02-11 UNCG-DAISY/psi-collect#85 @ebgoldstein
implement `click` package for path wrangling
>Using the click package can help tidy up some of the path wrangling that's being done in the pstorm.py file. Otherwise perhaps look at the argparse add_subparsers command. I'm not sure that using exec() to call the script is the best idea.

from @chrisleaman 's JOSS review: https://github.com/openjournals/joss-reviews/issues/2075#issuecomment-584910432
