### 2020-02-11 UNCG-DAISY/psi-collect#86 @ebgoldstein
reduce image size in test folder
>Try reducing the size data in the tests folder. Do you really need to include the full resolution images to test on? Or is it possible to use a reduced size image or even mock them out completely. This might improve compatibility with build systems (which can have issues if the repo is too big) and might just generally be easier to work with.

from @chrisleaman 's JOSS review: openjournals/joss-reviews#2075 (comment)
