### 2020-02-11 UNCG-DAISY/psi-collect#87 @ebgoldstein
use `pathlib` vs `os.path`
>I think a lot of the Path manipulation can be done with the pathlib library that was introduced in Python 3.4. This might help clean up the code, rather than using os.path.

from @chrisleaman 's JOSS review: openjournals/joss-reviews#2075
