### 2020-02-11 UNCG-DAISY/psi-collect#87 @ebgoldstein
`pathlib`: https://docs.python.org/3/library/pathlib.html
