### 2020-02-12 e585580
Bump mkdocs-minify-plugin from 0.2.1 to 0.2.2
Bumps [mkdocs-minify-plugin](https://github.com/byrnereese/mkdocs-minify-plugin) from 0.2.1 to 0.2.2.
- [Release notes](https://github.com/byrnereese/mkdocs-minify-plugin/releases)
- [Commits](https://github.com/byrnereese/mkdocs-minify-plugin/commits/0.2.2)

Signed-off-by: dependabot-preview[bot] <support@dependabot.com>
