### 2020-02-13 UNCG-DAISY/psi-collect#87 @Matmorcat
@chrisleaman I'm not aware of any particular performance improvements with our code using a high level library for path manipulation over the `os.path` module other than enforcing pure paths. What were your thoughts as far as what it might improve?
