### 2020-02-13 UNCG-DAISY/psi-collect#86 @Matmorcat
Yes. Some of the earlier tests were fairly inefficient, and I'd like to rewrite many of them as mocking out the actual accessing of web content and files would make the tests much more robust and reliable. For the current tests, the images can actually just be near-empty files, so long as the tests that check file sizes are updated as well.
