### 2020-02-13 UNCG-DAISY/psi-collect#87 @chrisleaman
From a performance perspective, I don't think there are any advantages of `pathlib` over `os.path`. But in terms of code and readability, I think [pathlib can make your code a bit cleaner](https://treyhunner.com/2018/12/why-you-should-be-using-pathlib/). I think one of the best things for this project would be the `glob` command, avoiding `os.walk` pattern and filtering out results:

```python
from pathlib import Path

top_level_csv_files = Path.cwd().glob('*.csv')
all_csv_files = Path.cwd().rglob('*.csv')
```

I do realize it is a bit of a pain to go back and refactor everything as the package works perfectly fine as it is. Sorry if this came across as _"you must use this package rather than that package"_ 😅! My suggestion was more changes/additions in the future and even for new other packages you might write 😃
