### 2020-02-17 b673343
Minor changes to paper.md (mostly gramatical)
