+++ paper/paper.md
@@ -46,21 +46,21 @@ storm impacts along developed and undeveloped coastlines [@Madore2018].
 
 Post-event imagery is typically large, both in terms of the number of individual image files and the size of each file. For 
 example, Hurricane Florence (2018) has over 29,000 JPEG images, with an average size of 7.7 Mb. The first steps for extracting 
-information from this data requires acquiring and processing images. NOAA ERI is currently available as a Web Map 
-Tile Service or via download using a graphical user interface (directing users to the relevant tar and ZIP archive). To enable 
+information from these data involve acquiring and processing images. NOAA ERI is currently available as a Web Map 
+Tile Service or via download using a graphical user interface (directing users to the relevant tar and ZIP archive) [Add reference to website]. To enable 
 users to download NOAA ERI images via command line for use in reproducible computational workflows, we developed a Python 
 module (`psi-collect`). 
 
 The key functionality of `psi-collect` is it allows users to download specific tar and ZIP archives based on storm name,
-date of image acquisition (day, month, and/or year), image type (JPEG, TIFF). Users can also filter and select specific 
+date of image acquisition (day, month, and/or year), and/or image type (JPEG, TIFF). Users can also filter and select specific 
 files using regular expressions. The module includes an automatic resumption feature in the event that a download 
 is interrupted. Each tar and ZIP archive is checked for integrity of contents upon download completion to ensure that 
-data is accurate and intact.
+data are accurate and intact.
 
 The module also functions as a tool for managing a user’s library of images — users can quickly understand which storms
 they have downloaded. A cataloging tool is also supplied, which allows users to build CSV files that display key
 information for each image such as image name, acquisition data, file size, and latitude and longitude for each corner
-of image (extracted from the associated `.geom` file). This catalog can be used for statistical and spatial analysis. 
+of image (extracted from the associated `.geom` file). This catalog can be used for statistical and spatial analyses. 
 
 `psi-collect` addresses four issues experienced by researchers working with large collections of NOAA post storm aerial imagery. 
 
@@ -71,7 +71,7 @@ individual web addresses for each storm archive (e.g., there are 15 individual a
 dynamically traverses pages and downloads archives when given specific filtering/sub-setting criteria (i.e., downloading all 
 archives from Hurricane Florence).
 
-Second, the individual image archives are large, and often freeze or terminate during download. Some browsers or software 
+Second, the individual image archives are large, and downloads often freeze or terminate early. Some browsers or software 
 tools may be capable of resuming partial downloads, but this is case specific and depends on the retrieval workflow. 
 Terminated downloads may require users to determine the missing archive, and re-initiate a download. `psi-collect` implements 
 both automatic resumption and post-download integrity checks to manage download interruption.
@@ -90,7 +90,7 @@ downloading is complete (and the archive is fully uploaded). This file is update
 `psi-collect` enables scientists to download NOAA Emergency Response Imagery via Python in a variety of ways (via date, 
 storm),and obtain metrics on downloaded images though the cataloging functions. We envision that `psi-collect` could be 
 used to develop reproducible computational workflows to analyze post-event imagery. For example, images can be used to: assess 
-damage to the built environment [e.g., @Thomas2014], measure impact using the @Sallenger2000 Storm Impact Scale [e.g., 
+damage to the built environment [e.g., @Thomas2014], measure impact in the context of predicted/hindcast @Sallenger2000 Storm Impact Scale regime [e.g., 
 @Liu2014; @Goldstein2020], evaluate forecasts of storm impact [@Morgan2019], measure the morphology of storm deposits [e.g., 
 @Overbeck2015; @Lazarus2016], and study how human development controls the shape of sediment and debris deposits [e.g., 
 @Rogers2015].
