+++ docs/requirements-docs.txt
@@ -2,6 +2,6 @@ markdown==3.2
 mkdocs==1.0.4
 mkdocs-git-revision-date-plugin==0.2
 mkdocs-macros-plugin==0.2.4
-mkdocs-material==4.6.2
+mkdocs-material==4.6.3
 mkdocs-minify-plugin==0.2.1
 pymdown-extensions==6.3
