### 2020-02-17 bd1df15
Bump markdown from 3.2 to 3.2.1 (#91)
Bumps [markdown](https://github.com/Python-Markdown/markdown) from 3.2 to 3.2.1.
- [Release notes](https://github.com/Python-Markdown/markdown/releases)
- [Commits](https://github.com/Python-Markdown/markdown/compare/3.2...3.2.1)

Signed-off-by: dependabot-preview[bot] <support@dependabot.com>

Co-authored-by: kodiakhq[bot] <49736102+kodiakhq[bot]@users.noreply.github.com>
