### 2020-02-17 UNCG-DAISY/psi-collect#95 @conlin-matt
Quick improvement to docs for clarity
I have a small suggestion to help make the main [docs page](https://psi-collect.readthedocs.io/en/master/) more clear:

There is a brief "Usage" section on the main page as well as a more detailed "Usage" section that can be navigated to. This is true for "Development" as well. Because of this, I found myself assuming that clicking anything in the section list on the left side would just take me to different sections of the main page, rather than to more detailed descriptions of things on different pages. Perhaps I am the only one that might have this issue, but in my opinion this could be easily fixed by simply adding a line at the end of the "Usage" section on the main page such as "For a more detailed description and examples see the 'Usage section' " where a click on that would take the reader to the more detailed page. This would have helped my understanding of the layout a lot. You sort of do this already in the "Development" section which is actually how I discovered my confusion.

I know this may be a bit knit-picky, but after all this is the time to make this great tool as user friendly as possible.

Note: I am submitting this issue as part of the JOSS review process (openjournals/joss-reviews#2075)
