### 2020-02-17 UNCG-DAISY/psi-collect#95 @ebgoldstein
this is a good idea @conlin-matt  — I will do this now
