### 2020-02-17 06bddb0
added link to the docs in 'Usage'
