+++ README.md
@@ -34,9 +34,13 @@ https://psi-collect.readthedocs.io/en/master/)
 
 ## Usage
 
+Quick Start:
+
 1. Install Python 3.6, 3.7, or 3.8 [**(Download Here)**](https://www.python.org/downloads/)
 2. Run `pip3 install psi-collect` in your favorite terminal
 3. Run `pstorm collect -h` for help on collecting images or `pstorm catalog -h` for help on cataloging local archives
+
+For a more detailed description and examples see the 'Usage section' in the [**Documentation**](https://psi-collect.readthedocs.io/en/master/)
 
 ## Contributing / Developing
 
