### 2020-02-17 UNCG-DAISY/psi-collect#95 @ebgoldstein
added the line under Usage section in these commits: https://github.com/UNCG-DAISY/psi-collect/commit/06bddb070b724e1e563b84634682d7a505d68a35 through https://github.com/UNCG-DAISY/psi-collect/commit/31cfc89aa290c6759b3cf2de23565fc30b7ea9aa .
