+++ README.md
@@ -40,7 +40,7 @@ Quick Start:
 2. Run `pip3 install psi-collect` in your favorite terminal
 3. Run `pstorm collect -h` for help on collecting images or `pstorm catalog -h` for help on cataloging local archives
 
-For a more detailed description and examples see the 'Usage section' in the [**Documentation**](https://psi-collect.readthedocs.io/en/master/)
+For a more detailed description and examples see the 'Usage' section in the [**Documentation**](https://psi-collect.readthedocs.io/en/master/)
 
 ## Contributing / Developing
 
