### 2020-02-18 UNCG-DAISY/psi-collect#96 @conlin-matt
Cataloging function not working for multiple storms?
**Describe the bug**
The cataloging function doesn't seem to work for a number of storms, possibly because these storms don't ship with .geom files? I have had issues with Sandy, Maria, Irma, and Irene. I notice that none of these seem to come with .geom files for their images (whereas storms such as Dorian and Gordon do).

**To Reproduce**
1. pstorm collect -s Sandy -a oct31ajpeg (or Irma or Maria etc.) -d
   *Notice the lack of .geom files
2. pstorm catalog -p .../Sandy (or Irma or Maria etc.)

**Expected behavior**
The catalog csv file should be produced.

**Screenshots**
I get this error message when trying to catalog images for Irma
![Screen Shot 2020-02-18 at 9 55 35 AM](https://user-images.githubusercontent.com/32144021/74747389-e4492f80-5234-11ea-8ea0-684089294196.png)

**Additional context**
This seems similar to #74 . Am I missing something here, e.g. downloading the images incorrectly so that the .geom files don't ship?

Note: Submitting this issue as part of the JOSS review process (openjournals/joss-reviews#2075)
