### 2020-02-18 UNCG-DAISY/psi-collect#96 @ebgoldstein
yes, unfortunately those storms didn't ship with raw `jpg`s, and therefore they don't have `.geom` files. Sandy has `jpg`s, but they are jpg tiles, so don't have associated
 `.geom` file separately.
I will add a note in the docs that cataloging the coordinates only work when storms have `.geom` files. there has been change every few years in what products are distributed.
