### 2020-02-18 UNCG-DAISY/psi-collect#96 @conlin-matt
Ok, that's what I figured. I agree that having that note in the docs is important in order to be transparent about what the tool can do, so glad to see it added.
