### 2020-02-18 UNCG-DAISY/psi-collect#96 @ebgoldstein
i'm reopening this until i make the changes in the docs
