### 2020-02-19 UNCG-DAISY/psi-collect#99 @katmratliff
Add feedback when no storm search term identified
Add short print statement when a storm search term or pattern is not found

**Describe the solution you'd like**
example response:
```
> pstorm collect -s katherine
Download Status Report (February 19, 2020 at 08:04 AM) <-s katherine -t .* -p C:\Users\<USER>\psi\collect\data\archives>

<No storms matching search term detected>
```

This is consistent with the feedback given when no archive files are detected for a specific search term or pattern.
