### 2020-02-19 UNCG-DAISY/psi-collect#99 @katmratliff
Note: Submitting this issue as part of the JOSS review process (openjournals/joss-reviews#2075)
