### 2020-02-19 UNCG-DAISY/psi-collect#96 @katmratliff
@ebgoldstein I've also noticed that when running `pstorm catalog` on a folder without `.geom` files can generate a `.csv` file that has an incorrect storm name... e.g., running on a folder on my desktop creates a `Desktop.csv` catalog file that also has the `storm_id` labeled as `desktop`. I've tested in different locations and end up with other incorrect catalog/`storm_id` names. Guessing this is similar to the path recursion issues in #74 (and I am also running on Windows). Maybe best to terminate generating a `csv` if there are no `.geom` files?
