### 2020-02-19 UNCG-DAISY/psi-collect#96 @ebgoldstein
thx for checking this @katmratliff. i think your suggestion — to terminate `.csv` generation when no `.geom` files exist — is the best way forward here.
