### 2020-02-20 #2075 @conlin-matt
Hi @kbarnhart et al.,

I have finished my review as well. Agreeing with @chrisleaman here that this is a neat tool to help deal with a whole-lot of difficult-to-manage data. I especially find the ability to automatically extract geographic info from the photos to be a hugely useful and important feature. It's relatively easy to use, and, with a few tweaks to the documentation (outlined in my submitted Issues above), is well documented and transparent in its capabilities. The paper is clear as well. With @chrisleaman's tweaks to the code, I don't see anything that glaringly needs improvement in terms of efficiency.

Once the relatively minor improvements to the cataloging routine are made (see above), I think all will look good.

Thanks to the authors for creating this tool- I've definitely added it to my toolbox!
