### 2020-02-20 c61dfb8
Bump mkdocs-minify-plugin from 0.2.1 to 0.2.3
Bumps [mkdocs-minify-plugin](https://github.com/byrnereese/mkdocs-minify-plugin) from 0.2.1 to 0.2.3.
- [Release notes](https://github.com/byrnereese/mkdocs-minify-plugin/releases)
- [Commits](https://github.com/byrnereese/mkdocs-minify-plugin/commits/0.2.3)

Signed-off-by: dependabot-preview[bot] <support@dependabot.com>
