+++ docs/requirements-docs.txt
@@ -3,5 +3,5 @@ mkdocs==1.0.4
 mkdocs-git-revision-date-plugin==0.2
 mkdocs-macros-plugin==0.2.4
 mkdocs-material==4.6.3
-mkdocs-minify-plugin==0.2.1
+mkdocs-minify-plugin==0.2.3
 pymdown-extensions==6.3
