### 2020-02-20 UNCG-DAISY/psi-collect#103 @katmratliff
problem extracting .zip on windows
**Describe the bug**
I get the `AttributeError` message `'ZipInfo' object has no attribute 'name'` when extracting .zip archives on my windows machine. This happens when downloading/extracting to the default directory or to a directory of my choice (e.g., on the desktop). Interestingly, the files do still seem to extract, but if you are trying to download and extract multiple archive files, this breaks the process so that only the first one gets downloaded and extracted.

**Screenshots**
If applicable, add screenshots to help explain your problem.
![image](https://user-images.githubusercontent.com/12630578/74982620-73f00900-5402-11ea-902f-67d99df6d775.png)

**Additional context**
It is very possible that this is an issue with my network/machine (federal government). I suspect this is the case if other windows users/reviewers haven't had this issue. I have no trouble downloading/extracting .tar archives.

Note: Submitting this issue as part of the JOSS review process (openjournals/joss-reviews#2075)
