### 2020-02-20 UNCG-DAISY/psi-collect#102 @Matmorcat
It looks like you're running it in a Conda environment which likely has an
older version of pandas. Have you tried it with pipenv, making sure to
include the project directory in your PYTHONPATH environment variable?

On Thu, Feb 20, 2020, 4:52 PM Katherine Ratliff <notifications@github.com>
wrote:

> *Describe the bug*
> test_cataloging.py fails on Win32 Python 3.7.3
>
> *To Reproduce*
> run pytest in root directory produces this:
> [image: image]
> <https://user-images.githubusercontent.com/12630578/74981825-f7a8f600-5400-11ea-8b98-ff4a6b7dcd01.png>
>
> *Additional context*
> I noticed that Travis was only set up to run in a linux env, so feel free
> to close this issue if you aren't worried about testing on Windows, just
> wanted you all to be aware that I had tried to run them and what the result
> was.
>
> Note: Submitting this issue as part of the JOSS review process (
> openjournals/joss-reviews#2075
> <https://github.com/openjournals/joss-reviews/issues/2075>)
>
