### 2020-02-20 #2075 @katmratliff
Hi @kbarnhart and others,

I have also finished my review. I agree with @chrisleaman and @conlin-matt that is a very useful package for streamlining the downloading, extracting, and processing of NOAA imagery, which will undoubtedly be of interest to many academic, engineering, and government entities. The paper states the use cases for the software, and the documentation is informative.  I did have trouble extracting .zip archive files on my machine, but I suspect this is an issue with my EPA machine and/or network, particularly if the other reviewers didn't have trouble doing this on windows machines (I think I saw @chrisleaman was also running windows?). I opened the following issues and a PR as part of my review:

- https://github.com/UNCG-DAISY/psi-collect/issues/103
- https://github.com/UNCG-DAISY/psi-collect/issues/102
- https://github.com/UNCG-DAISY/psi-collect/issues/99
- https://github.com/UNCG-DAISY/psi-collect/pull/104

Thanks to the `psi-collect` team for their hard work!
