### 2020-02-20 UNCG-DAISY/psi-collect#102 @chrisleaman
@katmratliff, I was able to get `pytest` running in my Windows environment so I think you're experiencing some issue with your local environment? Perhaps check you're using the latest version of numpy (see [this issue](https://github.com/scikit-learn-contrib/hdbscan/issues/272)).
