### 2020-02-20 UNCG-DAISY/psi-collect#103 @conlin-matt
@katmratliff I did not have any issues when I tested this on both MacOS and Windows (though I only tested for Hurricane Sandy...)
