### 2020-02-20 UNCG-DAISY/psi-collect#103 @Matmorcat
I was able to reproduce her issue with that specific archive, though the others I've tested work.

Thanks for catching this, @katmratliff !

EDIT: The issue is fixed by PR #105
