### 2020-02-20 1489024
minor suggestions to paper.md (#104)
Add specificity about what is slow to download

Co-authored-by: Matmorcat <mcmoretz@uncg.edu>
