### 2020-02-20 UNCG-DAISY/psi-collect#102 @katmratliff
@Matmorcat yes I was testing in a conda env. If @chrisleaman had no issues running the tests on windows then I will close this - thanks!
