### 2020-02-20 6737864
Use storm path to find the archive path reliably
