### 2020-02-21 65ac487
Fix date logging sometimes being incorrect & improve output
