+++ psicollect/cataloging/make_catalog.py
@@ -198,11 +198,26 @@ class Cataloging:
             if 'date' in current_fields_needed:
                 dates: List[str] = list()
 
+                stat_count_missing_date: int = 0
+
                 for i in range(len(files)):
 
                     print(f'\rGetting date taken from file {i + 1} of {len(files)} ({round((i / len(files)) * 100, 2)}%) ' +
                           '.' * (math.floor(((i + 1) % 9) / 3) + 1), end=' ')
-                    dates.append(Cataloging._get_best_date(os.path.join(scope_path, files[i])))
+
+                    best_date: str or None = Cataloging._get_best_date(os.path.join(scope_path, files[i]))
+
+                    if best_date is None:
+                        # If no date can be found, leave the entry blank
+                        dates.append(np.nan)
+                        stat_count_missing_date += 1
+                    else:
+                        dates.append(Cataloging._get_best_date(os.path.join(scope_path, files[i])))
+
+                if stat_count_missing_date > 0:
+                    print(str(stat_count_missing_date) + ' images had unknown dates!')
+                else:
+                    print('DONE')
 
                 catalog['date'] = dates
                 flag_unsaved_changes = True
@@ -342,7 +357,7 @@ class Cataloging:
     @staticmethod
     def _get_best_date(file_path: Union[bytes, str],
                        debug: bool = s.DEFAULT_DEBUG,
-                       verbosity: int = s.DEFAULT_VERBOSITY) -> str:
+                       verbosity: int = s.DEFAULT_VERBOSITY) -> str or None:
 
         # Assume years can only be 2000 to 2099 (current unix time ends at 2038 anyways)
         pattern: Pattern = re.compile('[\\D]*(20\\d{2})(\\d{2})(\\d{2})\\D')
@@ -361,10 +376,11 @@ class Cataloging:
 
             return year + '/' + month + '/' + day
 
-        # If no date can be parsed from file path or file name, then fallback to timestamp (sometimes off by a day)
+        # If no date can be parsed from file path or file name, then leave blank
         else:
-            h.print_error('Could not find any date in ' + file_path + ' ... resorting to file modify time!')
-            return Cataloging._timestamp_to_utc(os.path.getmtime(file_path))
+            if debug is True:
+                h.print_error('Could not find any date in ' + file_path + ' ... leaving it blank!')
+            return None
 
     @staticmethod
     def _timestamp_to_utc(timestamp: str or int) -> str:
