### 2020-02-21 cd05d73
Improve logging of incomplete catalogs
