### 2020-02-21 UNCG-DAISY/psi-collect#96 @Matmorcat
In the original post, @conlin-matt mentioned some storms that did not match the way the program identifies archives. I've changed the archive search to look for the parent directory being a valid storm name instead of matching arbitrary characters (original had to match a `20` and a `_` anywhere in the archive name). This should resolve the issue!

Make sure to **delete the old catalog** for the storms that had issues, before running again with the new changes.
