### 2020-02-21 287e7da
Fix recognition of archives in the cataloging process (#106)
* Fix an exception with extracting some zips

* Use storm path to find the archive path reliably

* Fix date logging sometimes being incorrect & improve output

* Improve logging of incomplete catalogs
