+++ psicollect/cataloging/make_catalog.py
@@ -170,7 +170,8 @@ class Cataloging:
                 entry: dict = dict()
                 entry['file'] = files[i]
                 entry['storm_id'] = Cataloging._get_storm_from_path(os.path.join(scope_path, files[i])).lower()
-                entry['archive'] = Cataloging._get_archive_from_path(os.path.join(scope_path, files[i])).lower()
+                entry['archive'] = Cataloging._get_archive_from_path(scope_path=os.path.join(scope_path, files[i]),
+                                                                     storm_id=entry['storm_id']).lower()
                 entry['image'] = Cataloging._get_image_from_path(os.path.join(scope_path, files[i]))
                 entries.append(entry)
 
@@ -197,11 +198,26 @@ class Cataloging:
             if 'date' in current_fields_needed:
                 dates: List[str] = list()
 
+                stat_count_missing_date: int = 0
+
                 for i in range(len(files)):
 
                     print(f'\rGetting date taken from file {i + 1} of {len(files)} ({round((i / len(files)) * 100, 2)}%) ' +
                           '.' * (math.floor(((i + 1) % 9) / 3) + 1), end=' ')
-                    dates.append(Cataloging._get_best_date(os.path.join(scope_path, files[i])))
+
+                    best_date: str or None = Cataloging._get_best_date(os.path.join(scope_path, files[i]))
+
+                    if best_date is None:
+                        # If no date can be found, leave the entry blank
+                        dates.append(np.nan)
+                        stat_count_missing_date += 1
+                    else:
+                        dates.append(Cataloging._get_best_date(os.path.join(scope_path, files[i])))
+
+                if stat_count_missing_date > 0:
+                    print(str(stat_count_missing_date) + ' images had unknown dates!')
+                else:
+                    print('DONE')
 
                 catalog['date'] = dates
                 flag_unsaved_changes = True
@@ -256,6 +272,7 @@ class Cataloging:
                 flag_unsaved_changes = True
 
         stat_files_accessed: int = 0
+        stat_count_missing_geom: int = 0
 
         # For any remaining fields needed (i.e. ll_lat), look for them in the .geom files
         for i, row in catalog.iterrows():
@@ -294,8 +311,18 @@ class Cataloging:
                         scope_path, os.path.normpath(row['file'])), debug=debug, verbosity=verbosity)
                 stat_files_accessed += 1
 
-                if geom_data is not None:
+                if geom_data is None:
+                    # The geom file does not exist
 
+                    geom_data = dict()
+
+                    for field_id in row_fields_needed:
+                        # Since no .geom file was found, fill with nan values
+                        geom_data[field_id] = np.nan
+
+                    stat_count_missing_geom += 1
+
+                else:
                     # Store the values in the catalog's respective column by field name, in memory
                     for key, value in geom_data.items():
                         try:
@@ -322,6 +349,11 @@ class Cataloging:
                       ' .geom files accessed) ... ', end='')
                 Cataloging._force_save_catalog(catalog=catalog, scope_path=scope_path)
 
+        if stat_count_missing_geom > 0:
+            print(str(stat_count_missing_geom) + ' images were missing a .geom file!')
+        else:
+            print('DONE')
+
         if debug:
             print('\r')
 
@@ -341,7 +373,7 @@ class Cataloging:
     @staticmethod
     def _get_best_date(file_path: Union[bytes, str],
                        debug: bool = s.DEFAULT_DEBUG,
-                       verbosity: int = s.DEFAULT_VERBOSITY) -> str:
+                       verbosity: int = s.DEFAULT_VERBOSITY) -> str or None:
 
[…]
