### 2020-02-21 UNCG-DAISY/psi-collect#99 @Matmorcat
You're right! I've added a message now when no storms match the search term. Added in a direct commit to master.
