+++ HISTORY.md
@@ -14,3 +14,9 @@
 
 - Fix issue with some older storms showing no archives
 - Fix display of some paths in collector showing incorrect separator
+
+## v1.0.3
+
+- Improve reliability of recognition of archive directories
+- Improve output of cataloging and reduce repetitive messages
+- Add message to known when no storms match a search expression
