+++ psicollect/__init__.py
@@ -4,4 +4,4 @@
 
 __name__ = 'psicollect'
 __author__ = 'PSI'
-__version__ = '1.0.2'
+__version__ = '1.0.3'
