+++ mkdocs.yml
@@ -62,4 +62,4 @@ markdown_extensions:
     - pymdownx.tilde
 
 extra:
-    version: 1.0.2
+    version: 1.0.3
