### 2020-02-21 7b89c39
Bump version: 1.0.2 → 1.0.3 (#107)
* Note on Travis-CI automation of distributing

* Update HISTORY.md

* Bump version: 1.0.2 → 1.0.3
