### 2020-02-21 58e7ed1
Update location of psicollect module text
