### 2020-02-21 23a43da
updata catalog docs
updated to reflect scope: catalogs that contain Latitude and Longitude information can only be built with storms that have 
both `.jpg` and `.geom` files.
