+++ docs/cataloging.md
@@ -4,7 +4,8 @@
 
 Once data is downloaded and unpacked, the data can be cataloged (`pstorm catalog <args>`) in order to more quickly
 access key information such as GPS data for each image, file size, and date taken to help with statistical and
-spacial analysis.
+spatial analysis. Note that catalogs that contain Latitude and Longitude information can only be built with storms that have 
+both `.jpg` and `.geom` files (e.g., Florence). 
 
 
 ## Command-Line Quick Reference
