### 2020-02-21 UNCG-DAISY/psi-collect#96 @ebgoldstein
docs now state that Lat/Lon in catalog relies on `.jpg` and `.geom` : https://github.com/UNCG-DAISY/psi-collect/commit/23a43dad392ac9c78733ef39a6b0ba629e1612a8
