### 2020-02-21 #2075 @ebgoldstein
Thanks @chrisleaman, @conlin-matt, @katmratliff, and @kbarnhart for all of your the suggestions, PRs, issues, bug identifications, and comments. The docs and the paper now reflect the changes requested by reviewers. I think we (read: @Matmorcat) have squashed some nice bugs.

@chrisleaman 's suggestions for future enhancements have been added as issues, and we hope to integrate these with future iterations of the tool.
