### 2020-02-22 #2075 @kbarnhart
Thanks to reviewers @chrisleaman, @conlin-matt, @katmratliff for prompt and thoughtful reviews and to authors @ebgoldstein (and @Matmorcat) for addressing issues. I agree with the decision to archive @chrisleaman's recommendations for future enhancements and outstanding issues. I see that those are the only remaining open issues.

At this stage, I ask that all reviewers confirm their recommendation to accept this publication.

After reviewers have confirmed, I will do a final copy edit of the manuscript.
