### 2020-02-24 #2075 @kbarnhart
@ebgoldstein can you:

* Issue a new tagged release reflecting the latest changes to the software.
* Respond here with that version number.
* Create a zenodo (or similar) archive. Make sure that it has the same title and author list as the paper.
* Post the DOI of that archive here.
