### 2020-02-25 5d365d9
JOSS final version bump (1.0.3 → 1.0.4) (#111)
* Update HISTORY.md

* Bump version: 1.0.3 → 1.0.4
