+++ mkdocs.yml
@@ -62,4 +62,4 @@ markdown_extensions:
     - pymdownx.tilde
 
 extra:
-    version: 1.0.3
+    version: 1.0.4
