### 2020-02-25 #2075 @Matmorcat
@kbarnhart I've created an updated release tagged [v1.0.4](https://github.com/UNCG-DAISY/psi-collect/releases/tag/v1.0.4). I'll leave @ebgoldstein to create the archive! 😄
