### 2020-02-25 #2075 @kbarnhart
@Matmorcat looks like deleting the post doesn't work.

I had tried your approach of re-posting with a shorter date range, and that didn't work... Thanks for the thought though!
