### 2020-02-26 14e6eca
Add Zenodo badge, remove Dependabot for brevity
